requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aniso8601==9.0.1
asgiref==3.2.10
async-timeout==4.0.3
attrs==23.2.0
bcrypt==3.2.0
blinker==1.7.0
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorama==0.4.4
cramjam==2.8.2
cryptography==3.4.8
dataclasses==0.6
distlib==0.3.8
distro==1.7.0
Django==3.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
dnspython==2.6.1
fasteners==0.14.1
filelock==3.13.1
frozenlist==1.4.1
future==0.18.2
greenlet==3.0.3
gunicorn==21.2.0
httplib2==0.20.2
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
marshmallow==3.21.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.5
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==23.2
paramiko==2.9.3
passlib==1.7.4
pbr==6.0.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.0
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
PyJWT==1.7.1
pykerberos==1.2.4
pymacaroons==0.13.0
pymongo==4.6.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dateutil==2.8.1
python-dotenv==1.0.1
python-snappy==0.7.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
stevedore==5.2.0
typing_extensions==4.9.0
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==3.0.1
xdg==5
yarl==1.9.4
zipp==1.0.0
